modify download scripts for comfyui

+++ services/comfy-ppio/clone.sh
@@ -2,10 +2,12 @@
 
 set -Eeuox pipefail
 
-mkdir -p /data/config/auto/extensions/"$1"
-cd /data/config/auto/extensions/"$1"
-git init
-git remote add origin "$2"
-git fetch origin "$3" --depth=1
-git reset --hard "$3"
-rm -rf .git
+if [ ! -d "/data/config/comfy/custom_nodes/$1" ]; then
+  mkdir -p /data/config/comfy/custom_nodes/"$1"
+  cd /data/config/comfy/custom_nodes/"$1"
+  git init
+  git remote add origin "$2"
+  git fetch origin "$3" --depth=1
+  git reset --hard "$3"
+  rm -rf .git
+fi
